% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AAdata}
\alias{AAdata}
\title{Properties, scales and indices for the 20 naturally occurring amino acids}
\format{
A list as follows of biochemical properites
}
\usage{
AAdata
}
\description{
A list with a collection of properties, scales and indices for the 20 naturally occurring amino acids from various sources.
}
\keyword{datasets}
