% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_imgt_junction_gaps.R
\name{add_imgt_junction_gaps}
\alias{add_imgt_junction_gaps}
\title{Insert Length Adjustment Gaps in IMGT JUNCTION sequences}
\usage{
add_imgt_junction_gaps(x, remove_non_canonicals = FALSE)
}
\arguments{
\item{x}{A character vector of IMGT JUNCTION sequences defined by the
conserved 2nd-CYS at position 104 through the J-PHE or J-TRP at position
118.}

\item{remove_non_canonicals}{A boolean.
\itemize{
\item \code{TRUE}: removes non-canonical JUNCTION sequences that lack a 2nd-CYS
and/or the J-PHE/J-TRP and replaces them with \code{NA}.
\item \code{FALSE}: (the default) leaves non-canonical JUNCTION sequences as they
are.
}}
}
\value{
A character vector the same length as \code{x}.
}
\description{
Adjust IMGT JUNCTION sequences for length by inserting gaps according to the
IMGT unique numbering rules. Vectorised for JUNCTION sequences and their
gap-inserted, length-adjusted replacements.
}
\details{
IMGT JUNCTION sequences shorter than 15 amino acids are modified by gap
insertions, following the IMGT unique numbering rules for JUNCTION
sequences, to a total length of 15 positions. The gap insertion character
used is ".". Non-canonical sequences for which the conserved 2nd-CYS and or
J-PHE/J-TRP cannot be identified can either removed or left as is without
being adjusted. It is recommended to first filter these sequences out if
downstream analysis depends on comparing Junction sequences that have been
length adjusted using this function (see examples). IMGT unique numbering
for JUNCTION sequences accommodates sequences longer than 15 amino acids,
but that is not implemented here yet. JUNCTION sequences longer than 15
positions are left as is.
}
\examples{
# Simple case of 1 sequence:
add_imgt_junction_gaps("CASSSGTAPFSFW")

# Multiple sequences containing short, long, and non-canonical JUNCTIONS:
cdr3_seqs <- c(
  "CASSF",
  "CASSGEKLFF",
  "CASSKPDRGIYGYTF",
  "TGPLHF",
  "CASSQETRYDFLTIDTGGKKKNTEAFF"
)
add_imgt_junction_gaps(cdr3_seqs)

# Remove non-canonical JUNCTIONS:
add_imgt_junction_gaps(cdr3_seqs, remove_non_canonicals = TRUE)

# Input sequences must be amino acids:
try(
add_imgt_junction_gaps("ATCATATTATATATCG")
)

# Input sequences must be a character vector:
try(
add_imgt_junction_gaps(c(104:118))
)
}
