% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aminoacid_physiochemical_feats}
\alias{aminoacid_physiochemical_feats}
\title{Amino Acid Physiochemical Features}
\format{
R data object. A data.frame.
}
\usage{
aminoacid_physiochemical_feats
}
\description{
A lookup table of values needed for calculating TiRP scores.
}
\keyword{datasets}
