% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cfselo_seqs}
\alias{cfselo_seqs}
\title{CFSE Low Immunoseq data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1042 rows and 20 columns.
}
\usage{
cfselo_seqs
}
\description{
CFSE Low Immunoseq data
}
\keyword{datasets}
