% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_sequences.R
\name{collapse_sequences}
\alias{collapse_sequences}
\title{Collapse Unique Sequences}
\usage{
collapse_sequences(data, .group_by = NULL)
}
\arguments{
\item{data}{A list or a single data frame as returned by
\code{\link[=read_immunoseq]{read_immunoseq()}}.}

\item{.group_by}{Character. The combination of T cell receptor sequences and
V and J genes to use in order to group unique sequences. If \code{NULL}, the
default is to group unique sequences by CDR3 nucleotide sequence plus the
most resolved V and J gene calls. Allowed inputs to \code{.group_by} are:
\describe{
\item{"nt"}{CDR3 nucleotide sequence}
\item{"ntvj"}{\emph{Default}. CDR3 nucleotide sequence plus most resolved V and J gene calls}
\item{"aa"}{CDR3 amino acid sequence}
\item{"aavj"}{CDR3 amino acid sequence plus most resolved V and J gene calls}
}}
}
\value{
A list or a single data frame of Immunoseq data with
}
\description{
Collapse repertoires using various combinations of unique sequence
definitions.
}
\details{
Rearrangement sequences are grouped according to the user selected unique
sequence definition and returned as list columns. The returned data will
contain one row per unique sequence. A column \code{duplicates} is added to
show the number of rearrangement sequences that match to a single
unique sequence. This function is useful for minimizing PCR and sequencing
errors that result in artificially inflated unique sequence counts.
}
\examples{
collapse_sequences(cfselo_seqs)
}
\seealso{
Other Immunoseq: 
\code{\link{diversity_methods}()},
\code{\link{format_immunarch}()},
\code{\link{get_alloreactives}()},
\code{\link{read_immunoseq}()},
\code{\link{remove_contaminants}()},
\code{\link{repertoire_diversity}()}
}
\author{
Christopher Parks
}
\concept{Immunoseq}
