% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_methods.R
\name{diversity_methods}
\alias{diversity_methods}
\title{Diversity Calculation Methods}
\usage{
diversity_methods()
}
\description{
For use with the \code{\link[=repertoire_diversity]{repertoire_diversity()}} function.
}
\details{
These are the various methods of diversity calculation that can be used to
quantify diversity of T cell receptor sequence repertoires. These methods
must be passed as individual character strings to the \code{.method} argument in
the \code{\link[=repertoire_diversity]{repertoire_diversity()}} function.
}
