% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_imgt_headers.R
\name{fix_imgt_fasta_headers}
\alias{fix_imgt_fasta_headers}
\title{Simplify IMGT FASTA Headers}
\usage{
fix_imgt_fasta_headers(.file)
}
\arguments{
\item{.file}{A character string. File path to a FASTA formatted sequence file
containing IMGT reference sequences. Alternatively a character vector R
object generated by first reading in the FASTA with \code{readLines()}.}
}
\description{
Overwrites IMGT FASTA headers with the V, D, or J gene name.
}
\details{
IMGT FASTA headers contain many useful fields that tell you at a glance
information about whether a gene is functional, a pseudogene, or an ORF.
There is also information about the positions where genes start and end
within a given sequence, as well as a verbose name for each sequence. At
times it is better to work with FASTA headers with simple names, for
example, when using them programmatically. This function simplifies IMGT
reference sequence FASTA headers to a V, D, or J gene name only.
}
\examples{
\dontrun{fasta <- data(vgenes)
  new_fasta <- fix_imgt_fasta_headers(fasta)
  writeLines(new_fastq)}

}
\author{
Christopher Parks (caparks2@gmail.com)
}
