% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_alloreactives.R
\name{get_alloreactives}
\alias{get_alloreactives}
\title{Define Alloreactive TCR Sequences}
\usage{
get_alloreactives(
  .cfselo,
  .unstim,
  .frequency = 2e-05,
  .fold = 2,
  .group_by = NULL,
  .use_reads = NULL
)
}
\arguments{
\item{.cfselo}{An input data frame (or list of data frames) as generated by
\code{\link[=read_immunoseq]{read_immunoseq()}} corresponding to the MLR responder T cells,
usually FACS sorted on the basis of being CFSE low. If there are separate
files for CD4 and CD8 responders \code{.cfselo} may also be a named list of such
data frames with format:
\verb{list(CD4 = CD4 data frame, CD8 = CD8 data frame)}.}

\item{.unstim}{An input data frame (or list of data frames) as generated by
\code{\link[=read_immunoseq]{read_immunoseq()}} corresponding to unstimulated T cells. If
there are separate files for CD4 and CD8 unstimulated T cells \code{.unstim} may
also be a named list of such data frames with format:
\verb{list(CD4 = CD4 data frame, CD8 = CD8 data frame)}.}

\item{.frequency}{Numeric. The repertoire template (or read) frequency
threshold for which responder T cells must be greater than or equal to in
order to be considered alloreactive. User may enter any numeric value, but
please note our convention is to use 1e-5 for reads data and
2e-5 for templates data due to copy number thresholds used by Immunoseq.}

\item{.fold}{Numeric. The multiplier defining the threshold for which
responder T cells must be greater than or equal to the template (or read)
frequency of a matching unstimulated T cells in order to be considered
alloreactive. User may enter any numeric value, but please note our
convention is to use 2.}

\item{.group_by}{Character. The combination of T cell receptor sequences and
V and J genes to use in order to group unique sequences. If \code{NULL}, the
default is to group unique sequences by CDR3 nucleotide sequence plus the
most resolved V and J gene calls. Allowed inputs to \code{.group_by} are:
\describe{
\item{"nt"}{CDR3 nucleotide sequence}
\item{"ntvj"}{\emph{Default}. CDR3 nucleotide sequence plus most resolved V and J gene calls}
\item{"aa"}{CDR3 amino acid sequence}
\item{"aavj"}{CDR3 amino acid sequence plus most resolved V and J gene calls}
\item{"rearrangement"}{full nucleotide sequence (not suitable for comparisons across platforms)}
\item{"extended_rearrangement"}{extended full nucleotide sequence (not suitable for comparisons across platforms)}
\item{"rearrangement_trunc"}{truncated full nucleotide sequence}
}}

\item{.use_reads}{Logical. \code{FALSE} (the default) indicates that repertoire
frequencies should be calculated using templates. \code{TRUE} indicates that
repertoire frequencies should be calculated using reads. Immunoseq files
are sometimes exported without reads information if templates are present
in the data. In this case \code{cdr3tools::get_alloreactives()} will throw an
error if \code{.use_reads} = TRUE.}
}
\value{
A data frame of alloreactive sequences. Columns defining unique
alloreactive sequences are determined by the user set value to \code{.group_by}.
Columns indicating the repertoire frequencies of each unique alloreactive
sequence in the responding and unstimulated repertoires are also included.
}
\description{
Determine alloreactive TCR sequences from Immunoseq sequencing data of
responding (CFSE Lo) T cells from an MLR and unstimulated T cells.
}
\examples{
get_alloreactives(.cfselo = cfselo_seqs, .unstim = unstim_seqs)
}
\references{
Obradovic A, Shen Y, Sykes M, Fu J. Integrated Analysis Toolset for Defining
and Tracking Alloreactive T-cell Clones After Human Solid Organ and
Hematopoietic Stem Cell Transplantation. Softw Impacts. 2021 Nov;10:100142.
doi: 10.1016/j.simpa.2021.100142. Epub 2021 Sep 23. PMID: 35291378; PMCID:
PMC8920412.
}
\seealso{
Other Immunoseq: 
\code{\link{diversity_methods}()},
\code{\link{read_immunoseq}()},
\code{\link{repertoire_diversity}()}
}
\author{
Christopher Parks
}
\concept{Immunoseq}
