% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{heldout_means_sds}
\alias{heldout_means_sds}
\title{Heldout Means SDS}
\format{
R data object. A data.frame.
}
\usage{
heldout_means_sds
}
\description{
A lookup table of values needed for calculating TiRP scores.
}
\keyword{datasets}
