% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrophobicity.R
\name{hydrophobicity}
\alias{hydrophobicity}
\title{Compute the hydrophobicity index of a protein sequence}
\source{
\url{https://github.com/dosorio/Peptides}
}
\usage{
hydrophobicity(seq, scale = "Wimley")
}
\arguments{
\item{seq}{An amino-acids sequence}

\item{scale}{A character string specifying the hydophobicity scale to be
used; must be one of \code{"Aboderin"}, \code{"AbrahamLeo"}, \code{"Argos"},
\code{"BlackMould"}, \code{"BullBreese"}, \code{"Casari"}, \code{"Chothia"},
\code{"Cid"}, \code{"Cowan3.4"}, \code{"Cowan7.5"}, \code{"Eisenberg"},
\code{"Engelman"}, \code{"Fasman"}, \code{"Fauchere"}, \code{"Goldsack"},
\code{"Guy"}, \code{"HoppWoods"}, \code{"Janin"}, \code{"Jones"},
\code{"Juretic"}, \code{"Kidera"}, \code{"Kuhn"}, \code{"KyteDoolittle"},
\code{"Levitt"}, \code{"Manavalan"}, \code{"Miyazawa"}, \code{"Parker"},
\code{"Ponnuswamy"}, \code{"Prabhakaran"}, \code{"Rao"}, \code{"Rose"},
\code{"Roseman"}, \code{"Sweet"}, \code{"Tanford"}, \code{"Welling"},
\code{"Wilson"}, \code{"Wolfenden"}, \code{"Zimmerman"},
\code{"interfaceScale_pH8"}, \code{"interfaceScale_pH2"},
\code{"octanolScale_pH8"}, \code{"octanolScale_pH2"}, \code{"oiScale_pH8"},
\code{"oiScale_pH2"}, or \code{"Wimley"}.}
}
\value{
The computed GRAVY index for a given amino-acid sequence
}
\description{
This function calculates the GRAVY hydrophobicity index of an amino acids
sequence using one of the 38 scales from different sources.
}
\details{
The hydrophobicity is an important stabilization force in protein folding;
this force changes depending on the solvent in which the protein is found.
The hydrophobicity index is calculated adding the hydrophobicity of
individual amino acids and dividing this value by the length of the
sequence.
}
\examples{
hydrophobicity(seq = "QWGRRCCGWGPGRRYCVRWC", scale = "Wimley")
}
