% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imgt_add_junction_gaps.R
\name{imgt_align_junctions}
\alias{imgt_align_junctions}
\title{Insert Alignment Gaps in IMGT JUNCTION sequences}
\usage{
imgt_align_junctions(.x, .rm_non_canonicals = FALSE)
}
\arguments{
\item{.x}{A character vector of IMGT JUNCTION sequences defined by the
conserved 2nd-CYS at position 104 through the J-PHE or J-TRP at position
118.}

\item{.rm_non_canonicals}{Logical.
\itemize{
\item \code{TRUE}: removes non-canonical JUNCTION sequences that lack a 2nd-CYS
and/or the J-PHE/J-TRP and replaces them with \code{NA}.
\item \code{FALSE}: (the default) leaves non-canonical JUNCTION sequences as they
are.
}}
}
\value{
A character vector the same length as \code{.x}.
}
\description{
Adjust IMGT JUNCTION sequences for length by inserting gaps according to the
IMGT unique numbering rules. Vectorised for JUNCTION sequences and their
gap-inserted, length-adjusted replacements.
}
\details{
IMGT JUNCTION sequences are modified by gap insertions, following the IMGT
unique numbering rules for JUNCTION sequences, up to the maximum length of
the longest sequence. The gap insertion character used is ".".
Non-canonical sequences for which the conserved 2nd-CYS and or J-PHE/J-TRP
cannot be identified can either removed or left as is.
}
\examples{
imgt_align_junctions(tcr_seqs$CDR3.aa)

imgt_align_junctions(c(tcr_seqs$CDR3.aa, "LRQQLEREGRYNEQFF"))

imgt_align_junctions(c(tcr_seqs$CDR3.aa, "LRQQLEREGRYNEQFF"),
                     .rm_non_canonicals = TRUE)

try(
  imgt_align_junctions("ATCATATTATATATCG")
)

try(
  imgt_align_junctions(c(104:118))
)

}
\references{
https://www.imgt.org/IMGTScientificChart/Numbering/IMGTIGVLsuperfamily.html

Lefranc M.-P., "Unique database numbering system for immunogenetic analysis",
Immunology Today, 18, 509 (1997). PMID: 9386342

Lefranc M.-P., "The IMGT unique numbering for Immunoglobulins, T cell
receptors and Ig-like domains", The Immunologist, 7, 132-136 (1999).

Lefranc, M.-P., Pommié, C., Ruiz, M., Giudicelli, V., Foulquier, E., Truong,
L., Thouvenin-Contet, V. and Lefranc, G., "IMGT unique numbering for
immunoglobulin and T cell receptor variable domains and Ig superfamily
V-like domains", Dev. Comp. Immunol., 27, 55-77 (2003). PMID: 12477501

Ruiz, M. and Lefranc, M.-P. "IMGT gene identification and Colliers de Perles
of human immunoglobulin with known 3D structures", Immunogenetics, 53,
857-883 (2002). PMID: 11862387

Kaas, Q. and Lefranc, M.-P. "IMGT Colliers de Perles: standardized
sequence-structure representations of the IgSF and MhcSF superfamily
domains", Current Bioinformatics, 2, 21-30 (2007).

Kaas, Q., Ruiz, M. and Lefranc, M.-P. "IMGT/3Dstructure-DB and
IMGT/StructuralQuery, a database and a tool for immunoglobulin, T cell
receptor and MHC structural data", Nucl. Acids. Res., 32, D208-D210 (2004).
PMID: 14681396
}
\seealso{
Other IMGT: 
\code{\link{imgt_format_gene_names}()},
\code{\link{imgt_get_ref_seqs}()},
\code{\link{imgt_simple_headers}()}
}
\author{
Christopher Parks
}
\concept{IMGT}
