% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imgt_extract.R
\name{imgt_extract}
\alias{imgt_extract}
\title{Extract Subregions of JUNCTION sequences}
\usage{
imgt_extract(
  .x,
  .start = 108,
  .end = 112,
  .rm_non_canonicals = FALSE,
  .aligned = FALSE
)
}
\arguments{
\item{.x}{A character vector of IMGT JUNCTION sequences defined by the
conserved 2nd-CYS at position 104 through the J-PHE or J-TRP at position
118. Alternatively, a data frame of TCR sequencing data as generated by
\link{read_immunoseq} or a list containing multiple of such data frames or
multiple of such character vectors.}

\item{.start}{An integer in [104, 117]. The starting position marking the
beginning of the subregion to be extracted, including the starting
position.}

\item{.end}{An integer in [105, 118]. The ending position marking the end of
the subregion to be extracted, including the ending position.}

\item{.rm_non_canonicals}{A logical. \code{FALSE} (the default) keeps
non-canonical JUNCTION sequences. \code{TRUE} removes them and replaces with NA.}

\item{.aligned}{A logical. \code{FALSE} (the default) returns results without
alignment gaps. This is useful for downstream analysis. \code{TRUE} returns
results with alignment gaps.}
}
\value{
A character vector the same length as \code{.x} with extracted JUNCTION
subregions, either with alignment gaps or not. Returns a list when \code{.x} is
a valid list. Returns a data frame with \code{.x} is a valid data frame.
}
\description{
Extract subregions from IMGT JUNCTION sequences in a manner consistent with
IMGT unique numbering rules.
}
\details{
By enforcing the IMGT unique numbering rules on JUNCTION sequences before
subsetting them, it can be assured that subregions extracted from multiple
sequences will have consistent positioning in the 3D structures of
TCR:pMHC complexes, allowing for fair comparisons between them. Comparisons
of unaligned sequences will usually result in sequence bias toward germline
elements (the somatic portions of CDR3 sequences). It is required to subset
JUNCTION sequences using the IMGT positions between the conserved 2nd-CYS
at position 104 and the J-TRP/J-PHE at position 118. At this time selecting
using the added positions between 111 and 112 (for example 112.3) is not
supported. Selections spanning 111 and 112 will return all the added
positions between them.
}
\examples{
imgt_extract(tcr_seqs$CDR3.aa)
imgt_extract(tcr_seqs$CDR3.aa, .aligned = TRUE)
}
\references{
https://www.imgt.org/IMGTScientificChart/Numbering/IMGTIGVLsuperfamily.html

Lefranc M.-P., "Unique database numbering system for immunogenetic analysis",
Immunology Today, 18, 509 (1997). PMID: 9386342

Lefranc M.-P., "The IMGT unique numbering for Immunoglobulins, T cell
receptors and Ig-like domains", The Immunologist, 7, 132-136 (1999).

Lefranc, M.-P., Pommié, C., Ruiz, M., Giudicelli, V., Foulquier, E., Truong,
L., Thouvenin-Contet, V. and Lefranc, G., "IMGT unique numbering for
immunoglobulin and T cell receptor variable domains and Ig superfamily
V-like domains", Dev. Comp. Immunol., 27, 55-77 (2003). PMID: 12477501

Ruiz, M. and Lefranc, M.-P. "IMGT gene identification and Colliers de Perles
of human immunoglobulin with known 3D structures", Immunogenetics, 53,
857-883 (2002). PMID: 11862387

Kaas, Q. and Lefranc, M.-P. "IMGT Colliers de Perles: standardized
sequence-structure representations of the IgSF and MhcSF superfamily
domains", Current Bioinformatics, 2, 21-30 (2007).

Kaas, Q., Ruiz, M. and Lefranc, M.-P. "IMGT/3Dstructure-DB and
IMGT/StructuralQuery, a database and a tool for immunoglobulin, T cell
receptor and MHC structural data", Nucl. Acids. Res., 32, D208-D210 (2004).
PMID: 14681396
}
\seealso{
Other IMGT: 
\code{\link{imgt_align_junctions}()},
\code{\link{imgt_format_gene_names}()},
\code{\link{imgt_get_ref_seqs}()},
\code{\link{imgt_simple_headers}()}
}
\author{
Christopher Parks
}
\concept{IMGT}
