% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imgt_format_gene_names.R
\name{imgt_format_gene_names}
\alias{imgt_format_gene_names}
\title{Format TCR VDJ Gene Names with IMGT Rules}
\usage{
imgt_format_gene_names(.x, .alleles = FALSE)
}
\arguments{
\item{.x}{A character vector of TCR V, D, or J gene names.}

\item{.alleles}{A logical. \code{FALSE} (the default) indicates that allele
identifiers (trailing "*01") should be removed. \code{TRUE} keeps them.}
}
\value{
A character vector of same length as \code{.x}.
}
\description{
Reformats character vectors of TCR VDJ gene names to conform to IMGT rules.
}
\details{
Vectorised over the input gene names character vector using
stringr::str_replace_all
}
\examples{
imgt_format_gene_names(c("TCRBV02-01*01", "TCRBV04-03*01", "TCRBV30-01*01"))
}
\references{
Lefranc M.-P., "Unique database numbering system for immunogenetic analysis",
Immunology Today, 18, 509 (1997). PMID: 9386342

Lefranc M.-P., "The IMGT unique numbering for Immunoglobulins, T cell
receptors and Ig-like domains", The Immunologist, 7, 132-136 (1999).

Lefranc, M.-P., Pommié, C., Ruiz, M., Giudicelli, V., Foulquier, E., Truong,
L., Thouvenin-Contet, V. and Lefranc, G., "IMGT unique numbering for
immunoglobulin and T cell receptor variable domains and Ig superfamily
V-like domains", Dev. Comp. Immunol., 27, 55-77 (2003). PMID: 12477501

Ruiz, M. and Lefranc, M.-P. "IMGT gene identification and Colliers de Perles
of human immunoglobulin with known 3D structures", Immunogenetics, 53,
857-883 (2002). PMID: 11862387

Kaas, Q. and Lefranc, M.-P. "IMGT Colliers de Perles: standardized
sequence-structure representations of the IgSF and MhcSF superfamily
domains", Current Bioinformatics, 2, 21-30 (2007).

Kaas, Q., Ruiz, M. and Lefranc, M.-P. "IMGT/3Dstructure-DB and
IMGT/StructuralQuery, a database and a tool for immunoglobulin, T cell
receptor and MHC structural data", Nucl. Acids. Res., 32, D208-D210 (2004).
PMID: 14681396
}
\seealso{
Other IMGT: 
\code{\link{imgt_add_junction_gaps}()},
\code{\link{imgt_get_ref_seqs}()},
\code{\link{imgt_simple_headers}()}
}
\author{
Christopher Parks
}
\concept{IMGT}
