% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imgt_get_ref_seqs.R
\name{imgt_get_ref_seqs}
\alias{imgt_get_ref_seqs}
\title{Fetch IMGT reference sequences}
\usage{
imgt_get_ref_seqs(
  .file,
  .simplify_headers = TRUE,
  .include_species = .simplify_headers
)
}
\arguments{
\item{.file}{A character vector. File path or URL to a FASTA formatted
sequence file containing IMGT reference sequences.}

\item{.simplify_headers}{Logical. \code{TRUE} (the default) will reduce sequence
names from the full length IMGT FASTA header to the gene name. \code{FALSE}
leaves the sequence name as the full length IMGT FASTA header.}

\item{.include_species}{Logical. \code{TRUE} (the default) will include the
species name in the sequence name. \code{FALSE} leaves the species name out.}
}
\value{
A named character vector with length equal to the number of entries
in the input FASTA file.
}
\description{
Fetch IMGT reference sequences in FASTA format and convert to a named
character vector.
}
\details{
Separate sequence lines under each FASTA entry are concatenated together into
single complete sequences, one for each FASTA entry, with a name taken from
the entry's header.
}
\examples{
\dontshow{if (url_exists()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Fetch from IMGT and simplify headers
file_path <- paste0(
  "https://www.imgt.org/download/V-QUEST/IMGT_V-QUEST",
  "_reference_directory/Homo_sapiens/TR/TRBV.fasta"
)

imgt_get_ref_seqs(file_path)
\dontshow{\}) # examplesIf}
}
\references{
https://www.imgt.org/vquest/refseqh.html
}
\seealso{
{imgt_add_junction_gaps()} {imgt_simple_headers()}

Other IMGT: 
\code{\link{imgt_add_junction_gaps}()},
\code{\link{imgt_simple_headers}()}
}
\author{
Christopher Parks
}
\concept{IMGT}
