% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_imgt_headers.R
\name{imgt_simple_headers}
\alias{imgt_simple_headers}
\title{Simplify IMGT FASTA Headers}
\usage{
imgt_simple_headers(.x, .include_species = FALSE)
}
\arguments{
\item{.x}{A character vector of FASTA formatted IMGT reference sequences,
first generated by manually reading in the FASTA file
with \code{readLines()} or with \code{get_imgt_ref_seqs()}}

\item{.include_species}{Logical. Indicates whether to include the species
name in the simplified header output.}
}
\value{
A character vector of FASTA formatted IMGT reference seqeuences of
same length as \code{.x} with headers reduced to gene names (with optional
species).
}
\description{
Overwrites IMGT FASTA headers with the V, D, or J gene name.
}
\details{
IMGT FASTA headers contain many useful fields that tell you at a glance
information about whether a gene is functional, a pseudogene, or an ORF.
There is also information about the positions where genes start and end
within a given sequence, as well as a verbose name for each sequence. At
times it is helpful to work with FASTA headers with simple names, for
example, when using them programmatically. This function simplifies IMGT
reference sequence FASTA headers to a V, D, or J gene name only.
}
\examples{
# Verbose headers
head(vgenes)

# Simplified headers
imgt_simple_headers(head(vgenes))

# Simplified headers with species name
imgt_simple_headers(head(vgenes), .include_species = TRUE)

}
\references{
https://www.imgt.org/IMGTindex/Fasta.php
}
\seealso{
{get_imgt_ref_seqs()}
}
\author{
Christopher Parks (caparks2@gmail.com)
}
\concept{IMGT FASTA Reference Sequences}
