% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{immunoseq_export_v1_col_names}
\alias{immunoseq_export_v1_col_names}
\title{Immunoseq v1 Export Column Names}
\format{
R data object. A character vector.
}
\usage{
immunoseq_export_v1_col_names
}
\description{
A file containing the column names for immunoseq v1 export .tsv files
}
\keyword{internal}
