% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jensen-shannon.R
\name{jensen_shannon}
\alias{jensen_shannon}
\title{Jensen Shannon Divergence and Distance}
\usage{
jensen_shannon(p, q, base = NULL, distance = NULL)
}
\arguments{
\item{p}{Numeric. Vector of unique sequence counts or frequencies (normalized
counts). Corresponds to the "templates", "Counts", "frequency", or
"Proportion" columns in data returned by \link{read_immunoseq} or
\link{format_immunarch}.}

\item{q}{Numeric. Vector of unique sequence counts or frequencies (normalized
counts). Corresponds to the "templates", "Counts", "frequency", or
"Proportion" columns in data returned by \link{read_immunoseq} or
\link{format_immunarch}.}

\item{base}{Numeric. The logarithm base to use in the calculation. If \code{NULL},
the default is base 2. Note that base e and base 10 are also commonly
used.}

\item{distance}{Logical. If \code{NULL} or \code{FALSE} (the default), calculates
Jensen Shannon Divergence. \code{TRUE} calculates Jensen Shannon Distance, the
square root of Jensen Shannon Divergence.}
}
\value{
Numeric. The value of Jensen Shannon Divergence, or at user's
option, the value of Jensen Shannon Distance, both using the default
base 2 logarithm or user supplied logarithm base value.
}
\description{
Calculate Jensen Shannon Divergence or Jensen Shannon Distance between two
antigen receptor sequence repertoires.
}
\details{
The Jensen Shannon Divergence (JSD) is calculated for normalized
vectors of unique sequence counts, or frequencies. Returned values are
between 0 and 1, with 0 meaning identical repertoires and 1 meaning
complete divergence. JSD is an information theory measure of the divergence
of TCR repertoires. This is a symmetric value defined for any two
vectors of normalised unique sequence counts \emph{P} and \emph{Q} in:

\deqn{\mathrm{JSD}(P \parallel Q) = \frac{1}{2}D(P \parallel M) + \frac{1}{2}D(Q \parallel M)}

where D is the Kullback Leibler Divergence (also called the relative
entropy) and M is a vector of point-wise means between \emph{P} and \emph{Q}. The
code was adapted as closely as possible from the scipy versions given in
the references.

To reproduce the JSD method used by Boris Grinshpun (Yufeng Shen and Megan
Sykes Labs) in the DeWolf 2018 JCI Insight paper (see references),
set \code{distance} = \code{TRUE}.

\emph{A special note:} R recycles atomic vectors during arithmetic operations if
the vectors are of unequal lengths. This means JSD will be calculated
incorrectly in R if the two repertoires being compared have differing
numbers of unique sequences. JSD for antigen receptor repertoires should
not be calculated this way, so the function will exit with an error
message. To prevent this, make sure input vectors p and q are of equal
lengths, and crucially, that each pairwise element of p and q correspond
to strictly the same unique sequence

\code{p} and \code{q} input vectors can be integer counts or normalized count doubles
(frequencies).
}
\examples{
p = c(0.10, 0.40, 0.50)
q = c(0.80, 0.15, 0.05)

# Jensen Shannon Divergence using ln
jensen_shannon(p, q, base = exp(1))

# Jensen Shannon Divergence using log2
jensen_shannon(p, q)

# Jensen Shannon Distance using log2
jensen_shannon(p, q, base = 2, distance = TRUE)
}
\references{
DeWolf S, Grinshpun B, Savage T, Lau SP, Obradovic A, Shonts B, Yang S,
Morris H, Zuber J, Winchester R, Sykes M, Shen Y. Quantifying size and
diversity of the human T cell alloresponse. JCI Insight.
2018 Aug 9;3(15):e121256. doi: 10.1172/jci.insight.121256.
PMID: 30089728; PMCID: PMC6129121.

\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC6129121/}

\url{https://docs.scipy.org/doc/scipy/reference/generated/scipy.spatial.distance.jensenshannon.html}

\url{https://docs.scipy.org/doc/scipy/reference/generated/scipy.stats.entropy.html}

\url{https://docs.scipy.org/doc/scipy/reference/generated/scipy.special.rel_entr.html}
}
\author{
Christopher Parks
}
