% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_contaminants.R
\name{remove_contaminants}
\alias{remove_contaminants}
\title{Remove Contaminant Clones}
\usage{
remove_contaminants(data, seqs)
}
\arguments{
\item{data}{A list or a single data frame as returned by
\code{\link[=read_immunoseq]{read_immunoseq()}} or \link{format_immunarch}.}

\item{seqs}{Character vector of CDR3 nucleotide and amino acid sequences.}
}
\value{
A list or a single data frame of repertoire data in the same format
as \code{data}, but with contaminant sequence rows removed.
}
\description{
Remove contaminant TCR sequences from Immunoseq or Immunarch repertoire data.
}
\details{
Accepts amino acid sequences and nucleotide sequences corresponding to
Immunoseq export v1 CDR3_amino_acid and CDR3_rearrangement/rearrangement
columns and Immunarch CDR3.aa and CDR3.nt/Sequence columns.
}
\examples{
remove_contaminants(
  cfselo_seqs,
  c("CASRDRGSYGYTF", "TGTGCCAGCAGGGCGACTAGCGGGAGGGCCTACGAGCAGTACTTC")
)
}
\seealso{
Other Immunoseq: 
\code{\link{collapse_sequences}()},
\code{\link{diversity_methods}()},
\code{\link{format_immunarch}()},
\code{\link{get_alloreactives}()},
\code{\link{read_immunoseq}()},
\code{\link{repertoire_diversity}()}
}
\author{
Christopher Parks
}
\concept{Immunoseq}
