% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repertoire_diversity.R
\name{repertoire_diversity}
\alias{repertoire_diversity}
\title{Measures of Diversity}
\usage{
repertoire_diversity(
  .x,
  .col = NULL,
  .method = cdr3tools::diversity_methods(),
  .r = NULL
)
}
\arguments{
\item{.x}{A data frame of antigen receptor sequencing data, minimally
containing a column with counts. Counts may represent reads, copies, or
templates etc of unique sequences (also called clonotypes or
rearrangments). Alternatively, \code{.x} may be a list of such data frames, a
single numeric vector of counts, or a list of numeric vectors of counts.}

\item{.col}{A character string - the name of the column containing counts
if \code{.x} is a data frame (or a list of data frames). \code{.col} accepts one
of the following:
* "reads"
* "seq_reads"
* "copies"
* "templates"
* "clones"
* "counts"
* Defaults to "templates" if no user supplied value and \code{.x} is a data
frame or list of data frames.}

\item{.method}{A character string of one of the following options:
\itemize{
\item \code{"shannons.clonality"}
\item \code{"shannons.entropy"}
\item \code{"shannons.diversity"}
\item \code{"gini.simpson"}
\item \code{"simpsons.clonality"}
\item \code{"simpsons.index"}
\item \code{"simpsons.dominance"}
\item \code{"simpsons.equitability"}
\item \code{"r20"}
\item \code{"slope"}
See ?cdr3tools::diversity_methods
}

Default is to return all diversity measures if no user supplied value.}

\item{.r}{A numeric vector of length 1. The fraction of top unique sequences
that, in their sum, account for the given \code{.r} proportion of total copies
(or reads or templates). Only used when \code{.method = "r20"}. Default value is
0.2 if no user supplied value.}
}
\value{
A numeric vector of length 1, or a data frame of results.
}
\description{
Numeric measures of species diversity derived from information theory
and ecology as applied to quantifying diversity of antigen receptor
repertoires.
}
\details{
Shannon's entropy is used in it's original form with log2, as that seems to
be the most commonly found version in the literature. The result is termed
"H". The form used more often in ecology is called Shannon's diversity
index and uses the natural log. For natural log the result is termed "H'",
otherwise it it "H" for log2 and log10. Shannon's Clonality is given as
1 - Pielou's index, which is also 1 - Shannon's equitability. Values range
from 0-1 with 1 being completely clonal while 0 is completely even.
Simpson's diversity index D1 is given as the probability that two clones
(template sequences) taken at random represent the same clonotype (unique
sequence or rearrangement). It is not often used because the values are not
intuitive. The Gini-Simpson index, often referred to as 'Simpson's Index',
is more commonly used than Simpson's diversity index D1 because the values
are more intuitive. Simpson's dominance D2, which is the inverse of
Simpson's index, is also often referred to as 'Simpson's Index'. It's use
is also more common than Simpson's diversity index D1. Simpson's
equitability is also given. Simpson's clonality is included as a normalized
measure and is potentially more useful than the others as it is more
robust to differences in sample sizes. It's discovery was made before
Simpson's diversity index D1 in 1945 by the economist Albert O. Hirschman
but is attributed to Simpson.

R20 is a semi-parametric measure of diversity that was
invented by Boris Grinshpun while working in Yufeng Shen's lab, in
collaboration with Megan Sykes, at Columbia University. The measure
describes the cumulative frequency of the most dominant unique sequences (
clonotypes or rearrangements), sorted from most to least abundant, that
account for 20\% (or any other selected value) of the total clones (reads,
templates, or copies) in the repertoire.

The abundance plot slope estimates the absolute value of the exponent in the
equation describing the power-law (Pareto) distribution. T cell receptor
repertoires have been shown to follow this distribution, with the vast
majority of unique sequences being found with only one copy, and extremely
few unique sequences being found with many copies. The slope of the line
resulting from linear regression of the log transformed abundance and
frequency variables quantifies the diversity of the repertoire. A steep
slope (higher absolute value) indicates a more diverse repertoire, while a
shallow slope (lower absolute value) indicates a more clonal repertoire.

Normalization to frequencies is not needed
since this is performed by the functions during calculation of each
diversity measure.
}
\examples{
template_counts <- c(100, 8, 3, 2, rep(1, times = 1e4))
repertoire_diversity(.x = template_counts, .method = "r20", .r = 0.5)

}
\references{
DeWolf S, Grinshpun B, Savage T, Lau SP, Obradovic A, Shonts B, Yang S,
Morris H, Zuber J, Winchester R, Sykes M, Shen Y. Quantifying size and
diversity of the human T cell alloresponse. JCI Insight.
2018 Aug 9;3(15):e121256. doi: 10.1172/jci.insight.121256. PMID: 30089728;
PMCID: PMC6129121

Obradovic A, Shen Y, Sykes M, Fu J. Integrated Analysis Toolset for Defining
and Tracking Alloreactive T-cell Clones After Human Solid Organ and
Hematopoietic Stem Cell Transplantation. Softw Impacts. 2021 Nov;10:100142.
doi: 10.1016/j.simpa.2021.100142. Epub 2021 Sep 23. PMID: 35291378; PMCID:
PMC8920412.
}
\seealso{
Other Immunoseq: 
\code{\link{diversity_methods}()},
\code{\link{get_alloreactives}()},
\code{\link{read_immunoseq}()}
}
\author{
Boris Grinshpun

Aleksandar Obradovic

Christopher Parks
}
\concept{Immunoseq}
