% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_frequency.R
\name{scale_frequency_min}
\alias{scale_frequency_min}
\alias{scale_frequency_max}
\title{Define Frequency Limits for Plot Scales}
\usage{
scale_frequency_min(x)

scale_frequency_max(x)
}
\arguments{
\item{x}{A numeric vector of frequencies. Must be in the set 0-1.}
}
\value{
A numeric of length = 1 representing the minimum or maximum, rounded
either up or down.
}
\description{
Given a vector a frequencies, defines the minimum and maximum limits and
rounds (down for the minimum and up for the maximum).
}
\details{
Particularly useful in the context of ggplot2. For example:
\verb{scale_y_continuous(limits = c(scale_frequency_min(x), scale_frequency_max(x))}
}
\examples{
scale_frequency_min(cfselo_seqs$frequency)

scale_frequency_max(cfselo_seqs$frequency)
}
\author{
Christopher Parks
}
