% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TiRP_weights}
\alias{TiRP_weights}
\title{TiRP Weights}
\format{
R data object. A data.frame.
}
\usage{
TiRP_weights
}
\description{
A lookup table of values needed for calculating TiRP scores.
}
\keyword{datasets}
