% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{unstim_seqs}
\alias{unstim_seqs}
\title{Unstimulated Immunoseq data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1042 rows and 20 columns.
}
\usage{
unstim_seqs
}
\description{
Unstimulated Immunoseq data
}
\keyword{datasets}
