% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_url.R
\name{url_exists}
\alias{url_exists}
\title{Check if a URL works}
\usage{
url_exists(
  x = paste0("https://www.imgt.org/download/V-QUEST/IMGT_V-QUEST",
    "_reference_directory/Homo_sapiens/TR/TRBV.fasta"),
  non_2xx_return_value = FALSE,
  quiet = FALSE,
  ...
)
}
\arguments{
\item{x}{a single URL}

\item{non_2xx_return_value}{what to do if the site exists but the
HTTP status code is not in the \verb{2xx} range. Default is to return \code{FALSE}.}

\item{quiet}{if not \code{FALSE}, then every time the \code{non_2xx_return_value} condition
arises a warning message will be displayed. Default is \code{FALSE}.}

\item{...}{other params (\code{timeout()} would be a good one) passed directly
to \code{httr::HEAD()} and/or \code{httr::GET()}}
}
\description{
A Bob Rudis Function
}
\references{
https://stackoverflow.com/a/52915256/11417996
}
\author{
Bob Rudis (hrbrmstr)
}
