% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vgenes}
\alias{vgenes}
\title{IMGT Human TRBV Reference Sequences}
\format{
FASTA formatted text file.
}
\source{
\url{https://www.imgt.org/download/V-QUEST/IMGT_V-QUEST_reference_directory/Homo_sapiens/TR/TRBV.fasta}
}
\usage{
vgenes
}
\description{
A file containing the IMGT Human TRBV reference set
used in IMGT/V-QUEST.
}
\keyword{datasets}
