% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_sequences.R
\name{collapse_sequences}
\alias{collapse_sequences}
\title{Collapse Unique Sequences}
\arguments{
\item{.group_by}{Character. The combination of T cell receptor sequences and
V and J genes to use in order to group unique sequences. If \code{NULL}, the
default is to group unique sequences by CDR3 nucleotide sequence plus the
most resolved V and J gene calls. Allowed inputs to \code{.group_by} are:
\describe{
\item{"nt"}{CDR3 nucleotide sequence}
\item{"ntvj"}{\emph{Default}. CDR3 nucleotide sequence plus most resolved V and J gene calls}
\item{"aa"}{CDR3 amino acid sequence}
\item{"aavj"}{CDR3 amino acid sequence plus most resolved V and J gene calls}
\item{"rearrangement"}{full nucleotide sequence (not suitable for comparisons across platforms)}
\item{"extended_rearrangement"}{extended full nucleotide sequence (not suitable for comparisons across platforms)}
\item{"rearrangement_trunc"}{truncated full nucleotide sequence}
}}

\item{.use_reads}{Logical. \code{FALSE} (the default) indicates that repertoire
frequencies should be calculated using templates. \code{TRUE} indicates that
repertoire frequencies should be calculated using reads. Immunoseq files
are sometimes exported without reads information if templates are present
in the data. In this case `` will throw an
error if \code{.use_reads} = TRUE.}
}
\description{
Collapse repertoires using various combinations of unique sequence
definitions
}
