% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repertoire_diversity.R
\name{diversity_methods}
\alias{diversity_methods}
\title{Diversity Calculation Methods}
\usage{
diversity_methods()
}
\description{
For use with the \code{\link[=repertoire_diversity]{repertoire_diversity()}} function.
}
\details{
These are the various methods of diversity calculation that can be used to
quantify diversity of T cell receptor sequence repertoires. These methods
must be passed as individual character strings to the \code{.method} argument in
the \code{\link[=repertoire_diversity]{repertoire_diversity()}} function. Available options to
\code{.method} are:
\itemize{
\item \code{"shannons.clonality"}
\item \code{"shannons.entropy"}
\item \code{"shannons.diversity"}
\item \code{"gini.simpson"}
\item \code{"simpsons.clonality"}
\item \code{"simpsons.index"}
\item \code{"simpsons.dominance"}
\item \code{"simpsons.equitability"}
\item \code{"r20"}
\item \code{"slope"}
}
}
\examples{
diversity_methods()
}
\seealso{
Other Immunoseq: 
\code{\link{collapse_sequences}()},
\code{\link{format_immunarch}()},
\code{\link{get_alloreactives}()},
\code{\link{read_immunoseq}()},
\code{\link{remove_contaminants}()},
\code{\link{repertoire_diversity}()}
}
\author{
Christopher Parks
}
\concept{Immunoseq}
