% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_TiRP_scores.R
\name{get_TiRP_scores}
\alias{get_TiRP_scores}
\title{Calculate TiRP Scores}
\usage{
get_TiRP_scores(.data)
}
\arguments{
\item{.data}{A data frame (and/or a tibble). This is the input TCR sequence
data that will be used to calculate TiRP scores. The format must conform to
the following:
\describe{
\item{$V.name}{First column. A character vector of TRBV gene names}
\item{$CDR3.aa}{Second column. A character vector of IMGT JUNCTION
sequences (CDR3 amino acid sequences from 2nd-CYS p104 to J-TRP/J-PHE
p118)}
}}
}
\value{
A data frame of TiRP scores for each evaluated unique sequence.
}
\description{
A scoring system to quantify TCR-intrinsic regulatory potential (TiRP)
}
\details{
A special note should be made to consider that this function, as written by
the authors, evaluates only sequences that are between 12 and 17 amino
acids in length and that evaluation is performed on sequences as they are,
in their linear, native states without first being adjusted for length by
IMGT unique numbering rules for JUNCTIONS.
}
\examples{
# Confirm the proper format for the input data
head(tcr_seqs)

TiRP(head(tcr_seqs))
}
\references{
Lagattuta KA, Kang JB, Nathan A, Pauken KE, Jonsson AH, Rao DA, Sharpe AH,
Ishigaki K, Raychaudhuri S. Repertoire analyses reveal T cell antigen
receptor sequence features that influence T cell fate. Nat Immunol. 2022
Mar;23(3):446-457. doi: 10.1038/s41590-022-01129-x. Epub 2022 Feb 17. PMID:
35177831; PMCID: PMC8904286.

https://www.nature.com/articles/s41590-022-01129-x

https://github.com/immunogenomics/TiRP
}
\author{
Kaitlyn A. Lagattuta, Joyce B. Kang, Aparna Nathan (Center for Data
Sciences, Brigham and Women’s Hospital, Boston, MA, USA)
Christopher Parks (rewritten and improved for use in this package.)
}
