% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jensen-shannon.R
\name{jensen_shannon}
\alias{jensen_shannon}
\title{Jensen Shannon Divergence (or Distance)}
\usage{
jensen_shannon(p, q, base = NULL, distance = NULL)
}
\arguments{
\item{p}{Numeric vector of repertoire counts or frequencies (normalized
counts). Corresponds to the "templates", "Counts", "frequency", or
"Proportion" columns in data returned by \link{read_immunoseq} or
\link{format_immunarch}.}

\item{q}{Numeric vector of repertoire counts or frequencies (normalized
counts). Corresponds to the "templates", "Counts", "frequency", or
"Proportion" columns in data returned by \link{read_immunoseq} or
\link{format_immunarch}.}

\item{base}{Numeric. The logarithm base to use in the calculation. Default
is the natural logarithm base e. Note that log base 2 is commonly used.}

\item{distance}{Logical. \code{FALSE} (the default) calculates Jensen Shannon
Divergence. \code{TRUE} calculates Jensen Shannon Distance, the square root of
Jensen Shannon Divergence.}
}
\value{
Numeric. The value of Jensen Shannon Divergence, or at user's
option, the value of Jensen Shannon Distance, both using the default
natural logarithm or user supplied logarithm base value.
}
\description{
Calculate Jensen Shannon Divergence or Distance between two
antigen receptor sequence repertoires.
}
\details{
The Jensen Shannon Divergence (or Distance), JSD, is calculated for normalized
repertoire vectors of counts (or frequencies) using relative entropy
(the Kullback Leibler Divergence). Returned values are between 0 and 1,
with 0 meaning identical repertoires and 1 meaning complete divergence.
JSD is an information theory-based measure of the divergence of TCR
repertoires. This is a symmetric value defined for any two repertoires p
and q as JSD(p, q) = (D(p || m) + D(q || m)) / 2. The code was adapted
as closely as possible from the Python Scipy versions given in the
references.
}
\examples{
p = c(0.10, 0.40, 0.50)
q = c(0.80, 0.15, 0.05)
# Jensen Shannon Divergence using ln
jensen_shannon(p, q)
# Jensen Shannon Divergence using log2
jensen_shannon(p, q, base = 2)
# Jensen Shannon Distance using log2
jensen_shannon(p, q, base = 2, distance = TRUE)
}
\references{
DeWolf S, Grinshpun B, Savage T, Lau SP, Obradovic A, Shonts B, Yang S,
Morris H, Zuber J, Winchester R, Sykes M, Shen Y. Quantifying size and
diversity of the human T cell alloresponse. JCI Insight.
2018 Aug 9;3(15):e121256. doi: 10.1172/jci.insight.121256.
PMID: 30089728; PMCID: PMC6129121.
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC6129121/}
\url{https://docs.scipy.org/doc/scipy/reference/generated/scipy.spatial.distance.jensenshannon.html}
\url{https://docs.scipy.org/doc/scipy/reference/generated/scipy.stats.entropy.html}
\url{https://docs.scipy.org/doc/scipy/reference/generated/scipy.special.rel_entr.html}
}
\author{
Christopher Parks
}
